Improve Aerial downloader UX and structure
