+++ scripts/aerial_downloader/__init__.py
@@ -1,0 +1,2 @@
+"""Helpers for downloading macOS Aerial videos from the local manifest."""
+
